BUG: Correctly handle 2-norm proximal operator for complex numbers
For now the solution is (sadly) to raise an error if it is attempted to
use the 2-norm proximal operator with a non-tight matrix and complex
numbers involved.
Complex numbers work as long as the matrix is tight, but for non-tight
matrices I need to find an optimisation routine for the proximal
operator that handles complex numbers.